requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.13
django-bootstrap==0.2.4
django-bootstrap-form==3.4
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-ckeditor==6.4.0
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-filter==21.1
django-jinja==2.10.0
django-jinja2==0.1
django-js-asset==2.0.0
django-modeltranslation==0.17.7
django-timezone-field==4.2.3
dnspython==2.2.1
docopt==0.6.2
eventlet==0.33.0
flower==1.0.0
greenlet==1.1.2
humanize==4.1.0
idna==3.3
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
pipreqs==0.4.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarg==0.1.9
